requirements.txt
mysql-connector-python==8.2.0
pandas==2.1.4
matplotlib==3.8.2
plotly==5.17.0
flask==3.0.0
python-dotenv==1.0.0 